requirements.txt
# =============================================
# HelpKart AI Agent - Python Dependencies
# =============================================

# FastAPI & Web Framework
fastapi
uvicorn[standard]
websockets

# Database
supabase
redis

# AI/ML
sentence-transformers
torch
groq

# Utilities
python-dotenv
tenacity
pydantic
pydantic-settings

# Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Development (Optional)
pytest
pytest-asyncio
httpx